refactored forecast options into its own package
