--- forecast/util.go
@@ -1,315 +1,0 @@
-package forecast
-
-import (
-	"errors"
-	"fmt"
-	"log/slog"
-	"math"
-	"strconv"
-	"time"
-
-	"github.com/aouyang1/go-forecaster/feature"
-	"golang.org/x/exp/slices"
-	"gonum.org/v1/gonum/floats"
-)
-
-var ErrUnknownTimeFeature = errors.New("unknown time feature")
-
-const (
-	LabelTimeEpoch = "epoch"
-
-	LabelSeasDaily  = "daily"
-	LabelSeasWeekly = "weekly"
-
-	LabelEventWeekend = "weekend"
-)
-
-func generateTimeFeatures(t []time.Time, opt *Options) (*feature.Set, *feature.Set) {
-	if opt == nil {
-		opt = NewDefaultOptions()
-	}
-
-	tFeat := feature.NewSet()
-
-	epoch := make([]float64, len(t))
-	for i, tPnt := range t {
-		epochNano := float64(tPnt.UnixNano()) / 1e9
-		epoch[i] = epochNano
-	}
-	feat := feature.NewTime(LabelTimeEpoch)
-	tFeat.Set(feat, epoch)
-
-	eFeat := generateEventFeatures(t, opt)
-	tFeat.Update(eFeat)
-
-	return tFeat, eFeat
-}
-
-func generateEventFeatures(t []time.Time, opt *Options) *feature.Set {
-	winFunc := WindowFunc(opt.MaskWindow)
-
-	eFeat := feature.NewSet()
-
-	opt.WeekendOptions.generateEventMask(t, eFeat, winFunc)
-	opt.EventOptions.generateEventMask(t, eFeat, winFunc)
-	return eFeat
-}
-
-func generateEventMaskWithFunc(t []time.Time, maskCond func(tPnt time.Time) bool, windowFunc func(seq []float64) []float64) []float64 {
-	mask := make([]float64, len(t))
-	var maskSpans [][2]int
-	var inMask bool
-	var maskSpan [2]int
-	for i, tPnt := range t {
-		if maskCond(tPnt) {
-			if !inMask {
-				inMask = true
-				maskSpan[0] = i
-			}
-			mask[i] = 1.0
-			continue
-		}
-		if inMask {
-			inMask = false
-			maskSpan[1] = i
-			maskSpans = append(maskSpans, maskSpan)
-		}
-	}
-	if inMask {
-		maskSpan[1] = len(t)
-		maskSpans = append(maskSpans, maskSpan)
-	}
-
-	for _, maskSpan := range maskSpans {
-		windowFunc(mask[maskSpan[0]:maskSpan[1]])
-	}
-	return mask
-}
-
-func generateFourierFeatures(feat *feature.Set, opt *Options) (*feature.Set, error) {
-	if opt == nil {
-		opt = NewDefaultOptions()
-	}
-	x := feature.NewSet()
-
-	opt.SeasonalityOptions.removeDuplicates()
-
-	periods := make(map[float64]struct{})
-	colinearCfgOrders := make(map[SeasonalityConfig][]int)
-	for _, seasCfg := range opt.SeasonalityOptions.SeasonalityConfigs {
-		for i := 1; i <= seasCfg.Orders; i++ {
-			period := float64(seasCfg.Period) / float64(i)
-			if _, exists := periods[period]; exists {
-				// store colinear period
-				colinearCfgOrders[seasCfg] = append(colinearCfgOrders[seasCfg], i)
-				continue
-			}
-			periods[period] = struct{}{}
-		}
-	}
-
-	for _, seasCfg := range opt.SeasonalityOptions.SeasonalityConfigs {
-		var orders []int
-		for i := 1; i <= seasCfg.Orders; i++ {
-			colinearOrders, colinearCfgExists := colinearCfgOrders[seasCfg]
-			if colinearCfgExists && slices.Contains(colinearOrders, i) {
-				continue
-			}
-			orders = append(orders, i)
-		}
-		seasFeatures, err := generateFourierOrders(feat, orders, seasCfg.Period, seasCfg.Name)
-		if err != nil {
-			return nil, fmt.Errorf("unable to generate seasonality features for %q, %w", seasCfg.Name, err)
-		}
-		x.Update(seasFeatures)
-
-		switch seasCfg.Name {
-		case LabelSeasDaily:
-			// only model for daily since we're masking the weekends which means we do not meet the sampling requirements
-			// to capture weekly seasonality.
-			if opt.WeekendOptions.Enabled {
-				eventSeasFeat, err := generateEventSeasonality(feat, seasFeatures, LabelEventWeekend, LabelSeasDaily)
-				if err != nil {
-					slog.Warn("unable to generate weekend daily seasonality", "feature_name", LabelEventWeekend)
-				} else {
-					x.Update(eventSeasFeat)
-				}
-			}
-		}
-
-		for _, e := range opt.EventOptions.Events {
-			eventSeasFeat, err := generateEventSeasonality(feat, seasFeatures, e.Name, seasCfg.Name)
-			if err != nil {
-				slog.Warn("unable to generate event seasonality", "feature_name", e.Name, "seasonality", seasCfg.Name)
-				continue
-			}
-
-			x.Update(eventSeasFeat)
-		}
-
-	}
-	return x, nil
-}
-
-func generateFourierOrders(tFeatures *feature.Set, orders []int, periodDur time.Duration, label string) (*feature.Set, error) {
-	if tFeatures == nil {
-		return nil, ErrUnknownTimeFeature
-	}
-
-	col := LabelTimeEpoch
-	tFeat, exists := tFeatures.Get(feature.NewTime(col))
-	if !exists {
-		return nil, ErrUnknownTimeFeature
-	}
-
-	period := periodDur.Seconds()
-
-	x := feature.NewSet()
-	for _, order := range orders {
-		sinFeat, cosFeat := generateFourierComponent(tFeat, order, period)
-		sinFeatCol := feature.NewSeasonality(col+"_"+label, feature.FourierCompSin, order)
-		cosFeatCol := feature.NewSeasonality(col+"_"+label, feature.FourierCompCos, order)
-		x.Set(sinFeatCol, sinFeat)
-		x.Set(cosFeatCol, cosFeat)
-	}
-
-	return x, nil
-}
-
-func generateFourierComponent(timeFeature []float64, order int, period float64) ([]float64, []float64) {
-	omega := 2.0 * math.Pi * float64(order) / period
-	sinFeat := make([]float64, len(timeFeature))
-	cosFeat := make([]float64, len(timeFeature))
-	for i, tFeat := range timeFeature {
-		rad := omega * tFeat
-		sinFeat[i] = math.Sin(rad)
-		cosFeat[i] = math.Cos(rad)
-	}
-	return sinFeat, cosFeat
-}
-
-func generateEventSeasonality(feat, sFeat *feature.Set, eCol, sLabel string) (*feature.Set, error) {
-	mask, exists := feat.Get(feature.NewEvent(eCol))
-	if !exists {
-		return nil, fmt.Errorf("event mask not found, skipping event name, %s", eCol)
-	}
-
-	eventSeasonalityFeatures := feature.NewSet()
-	for _, label := range sFeat.Labels() {
-		featData, exists := sFeat.Get(label)
-		if !exists {
-			continue
-		}
-		maskedData := make([]float64, len(featData))
-		floats.MulTo(maskedData, mask, featData)
-
-		fcompStr, _ := label.Get("fourier_component")
-		fcomp := feature.FourierComp(fcompStr)
-
-		orderStr, _ := label.Get("order")
-		order, _ := strconv.Atoi(orderStr)
-		featCol := feature.NewSeasonality(eCol+"_"+sLabel, fcomp, order)
-		eventSeasonalityFeatures.Set(featCol, maskedData)
-	}
-	return eventSeasonalityFeatures, nil
-}
-
-func generateAutoChangepoints(t []time.Time, n int) []Changepoint {
-	var minTime, maxTime time.Time
-	for _, tPnt := range t {
-		if minTime.IsZero() || tPnt.Before(minTime) {
-			minTime = tPnt
-		}
-		if maxTime.IsZero() || tPnt.After(maxTime) {
-			maxTime = tPnt
-		}
-	}
-
-	window := maxTime.Sub(minTime)
-	changepointWinNs := int64(window.Nanoseconds()) / int64(n)
-	chpts := make([]Changepoint, 0, n)
-
-	for i := 0; i < n; i++ {
-		chpntTime := minTime.Add(time.Duration(changepointWinNs * int64(i)))
-		chpts = append(
-			chpts,
-			NewChangepoint("auto_"+strconv.Itoa(i), chpntTime),
-		)
-	}
-	return chpts
-}
-
-func generateChangepointFeatures(t []time.Time, chpts []Changepoint, trainingEndTime time.Time, enableGrowth bool) *feature.Set {
-	filteredChpts := make([]Changepoint, 0, len(chpts))
-	for _, chpt := range chpts {
-		// skip over changepoints that are after the training end time since they'll
-		// not have been modeled producing zeroes in the feature set
-		if chpt.T.After(trainingEndTime) {
-			continue
-		}
-		filteredChpts = append(filteredChpts, chpt)
-	}
-
-	// create a slice of features where it goes in the order of bias, slope for each changepoint
-	chptBiasFeatures := make([][]float64, len(filteredChpts))
-	var chptGrowthFeatures [][]float64
-	if enableGrowth {
-		chptGrowthFeatures = make([][]float64, len(filteredChpts))
-	}
-	for i := 0; i < len(filteredChpts); i++ {
-		chptBias := make([]float64, len(t))
-		chptBiasFeatures[i] = chptBias
-
-		if enableGrowth {
-			chptGrowth := make([]float64, len(t))
-			chptGrowthFeatures[i] = chptGrowth
-		}
-
-	}
-
-	// compute dt between training end time and changepoint time
-	deltaT := make([]float64, len(filteredChpts))
-	for i, chpt := range filteredChpts {
-		deltaT[i] = trainingEndTime.Sub(chpt.T).Seconds()
-	}
-
-	bias := 1.0
-	var slope float64
-	for i := 0; i < len(t); i++ {
-		for j := 0; j < len(filteredChpts); j++ {
-			if t[i].Equal(filteredChpts[j].T) || t[i].After(filteredChpts[j].T) {
-				chptBiasFeatures[j][i] = bias
-
-				if enableGrowth {
-					slope = t[i].Sub(filteredChpts[j].T).Seconds() / deltaT[j]
-					chptGrowthFeatures[j][i] = slope
-				}
-			}
-		}
-	}
-
-	feat := feature.NewSet()
-	for i := 0; i < len(filteredChpts); i++ {
-		chpntName := strconv.Itoa(i)
-		if filteredChpts[i].Name != "" {
-			chpntName = filteredChpts[i].Name
-		}
-		chpntBias := feature.NewChangepoint(chpntName, feature.ChangepointCompBias)
-		feat.Set(chpntBias, chptBiasFeatures[i])
-
-		if enableGrowth {
-			chpntGrowth := feature.NewChangepoint(chpntName, feature.ChangepointCompSlope)
-			feat.Set(chpntGrowth, chptGrowthFeatures[i])
-		}
-	}
-	return feat
-}
-
-func indentExpand(indent string, growth int) string {
-	indentByte := []byte(indent)
-	out := make([]byte, 0, len(indent)*growth)
-	for i := 0; i < growth; i++ {
-		out = append(out, indentByte...)
-	}
-	return string(out)
-}
